Conflicts:
	Code/assets/Configuration.xsd
	Code/logic/AgentConfig.java
	Code/logic/PowerUpConfig.java
	Code/logic/PredatorPowerType.java
	Code/logic/PreyPowerType.java
	Code/xml/PredatorPowerUpTypeAdapter.java
	Code/xml/PreyPowerUpAdapter.java
	Code/xml/PreyPowerUpTypeAdapter.java
	Code/xml/XmlPredatorPowerUp.java
	Code/xml/XmlPreyPowerUp.java

+++ Code/xml/XmlPoint.java
@@ -12,15 +12,15 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement(name = "XmlPoint")
 public class XmlPoint {
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	/**
 	 * Get the X coordinate.
 	 * 
 	 * @return the X coordinate.
 	 */
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 	
@@ -30,7 +30,7 @@ public class XmlPoint {
 	 * @param x - the value to set the X coordinate to.
 	 */
 	@XmlElement (name = "X")
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
@@ -39,7 +39,7 @@ public class XmlPoint {
 	 * 
 	 * @return the Y coordinate.
 	 */
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 	
@@ -49,7 +49,7 @@ public class XmlPoint {
 	 * @param y - the value to set the Y coordinate to.
 	 */
 	@XmlElement (name = "Y")
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
